smartnoise-synthetic/requirements.txt
opendp-smartnoise
diffprivlib
scikit-learn==0.23.2
pandas==0.25.3
pytest
pyyaml
pandasql
mlflow
connexion
ctgan==0.2.2
opacus==0.11.0
torch==1.6.0
seaborn
